update: rename new footer parser
